@@ -12,31 +12,46 @@
     "description": "A coroutine framework that focuses on hyperspeed and flexible, specifically use for build microservices and middlewares.",
     "license": "Apache-2.0",
     "require": {
-        "php": ">=7.3",
-        "hyperf/cache": "~2.2.0",
-        "hyperf/command": "~2.2.0",
-        "hyperf/config": "~2.2.0",
-        "hyperf/db-connection": "~2.2.0",
-        "hyperf/framework": "~2.2.0",
-        "hyperf/guzzle": "~2.2.0",
-        "hyperf/http-server": "~2.2.0",
-        "hyperf/logger": "~2.2.0",
-        "hyperf/memory": "~2.2.0",
-        "hyperf/process": "~2.2.0",
-        "hyperf/database": "~2.2.0",
-        "hyperf/redis": "~2.2.0",
-        "hyperf/constants": "~2.2.0",
-        "hyperf/async-queue": "~2.2.0",
-        "hyperf/model-cache": "~2.2.0",
-        "hyperf/elasticsearch": "~2.2.0"
+        "php": ">=8.0",
+        "ext-mongodb": "*",
+        "96qbhy/hyperf-auth": "^3.0",
+        "hyperf/async-queue": "~3.0.0",
+        "hyperf/cache": "~3.0.0",
+        "hyperf/code-generator": "^0.3.3",
+        "hyperf/command": "~3.0.0",
+        "hyperf/config": "~3.0.0",
+        "hyperf/constants": "~3.0.0",
+        "hyperf/database": "~3.0.0",
+        "hyperf/db-connection": "~3.0.0",
+        "hyperf/elasticsearch": "~3.0.0",
+        "hyperf/filesystem": "~3.0.0",
+        "hyperf/flysystem-oss": "^1.0",
+        "hyperf/framework": "~3.0.0",
+        "hyperf/guzzle": "~3.0.0",
+        "hyperf/http-server": "~3.0.0",
+        "hyperf/logger": "~3.0.0",
+        "hyperf/memory": "~3.0.0",
+        "hyperf/model-cache": "~3.0.0",
+        "hyperf/paginator": "~3.0.0",
+        "hyperf/process": "~3.0.0",
+        "hyperf/redis": "~3.0.0",
+        "hyperf/session": "~3.0.0",
+        "hyperf/snowflake": "~3.0.0",
+        "hyperf/socketio-server": "~3.0",
+        "hyperf/task": "~3.0",
+        "hyperf/validation": "~3.0",
+        "hyperf/websocket-server": "^3.0",
+        "overtrue/easy-sms": "^2.2",
+        "phper666/jwt-auth": "^4.0"
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^3.0",
-        "hyperf/devtool": "~2.2.0",
-        "hyperf/ide-helper": "~2.2.0",
-        "hyperf/testing": "~2.2.0",
+        "hyperf/devtool": "~3.0.0",
+        "hyperf/ide-helper": "~3.0.0",
+        "hyperf/testing": "~3.0.0",
+        "hyperf/watcher": "^3.0",
         "mockery/mockery": "^1.0",
-        "phpstan/phpstan": "^0.12",
+        "phpstan/phpstan": "^1.8",
         "swoole/ide-helper": "^4.5"
     },
     "suggest": {
