@@ -2,15 +2,13 @@
 
 declare(strict_types=1);
 /**
- * This file is part of Hyperf.
- *
- * @link     https://www.hyperf.io
- * @document https://hyperf.wiki
- * @contact  group@hyperf.io
- * @license  https://github.com/hyperf/hyperf/blob/master/LICENSE
+ * This file is part of Nursery2.
+ * @author    denglei@4587@163.com
  */
 namespace App\Controller;
 
+use App\Exception\BusinessException;
+use Hyperf\Contract\StdoutLoggerInterface;
 use Hyperf\Di\Annotation\Inject;
 use Hyperf\HttpServer\Contract\RequestInterface;
 use Hyperf\HttpServer\Contract\ResponseInterface;
@@ -18,21 +16,26 @@ use Psr\Container\ContainerInterface;
 
 abstract class AbstractController
 {
-    /**
-     * @Inject
-     * @var ContainerInterface
-     */
-    protected $container;
+    #[Inject]
+    protected ContainerInterface $container;
 
-    /**
-     * @Inject
-     * @var RequestInterface
-     */
-    protected $request;
+    #[Inject]
+    protected RequestInterface $request;
 
-    /**
-     * @Inject
-     * @var ResponseInterface
-     */
-    protected $response;
+    #[Inject]
+    protected ResponseInterface $response;
+
+    #[Inject]
+    protected StdoutLoggerInterface $logger;
+
+    public function success($data, $message = 'success')
+    {
+        $code = $this->response->getStatusCode();
+        return $this->response->json(['message' => $message, 'code' => $code, 'data' => $data]);
+    }
+
+    public function failed($message = '请求错误!', $code = 500, $data = [])
+    {
+        throw new BusinessException($code, $message);
+    }
 }
