@@ -1,6 +1,9 @@
 #!/usr/bin/env php
 <?php
-
+/**
+ * This file is part of Nursery2.
+ * @author    denglei@4587@163.com
+ */
 ini_set('display_errors', 'on');
 ini_set('display_startup_errors', 'on');
 ini_set('memory_limit', '1G');
