@@ -2,14 +2,15 @@
 
 declare(strict_types=1);
 /**
- * This file is part of Hyperf.
- *
- * @link     https://www.hyperf.io
- * @document https://hyperf.wiki
- * @contact  group@hyperf.io
- * @license  https://github.com/hyperf/hyperf/blob/master/LICENSE
+ * This file is part of Nursery2.
+ * @author    denglei@4587@163.com
  */
 return [
     'http' => [
+        # \Hyperf\Validation\Middleware\ValidationMiddleware::class
+        # \Hyperf\Session\Middleware\SessionMiddleware::class
+    ],
+    'socket-io' => [
+        # \App\Middleware\SocketJwtAuthMiddleware::class,
     ],
 ];
